redoing the code (work in progress)
What I'm trying to do right now is refactor everything, which I have mostly done.

But also I'm trying to find a way to search words that have a letter in a  specific position. Once I do that, then I will work on UI code so we can make this run in a simple little text box / user input kind of window. Then I'll be done with this project for good!