FINALLY FIXED
Now the code can take into account words with positional letters.

So now you can tell the script where a letter is in the word if you know it.

the only flaw left here is that if you have an extra (duplicate) letter in the word it won't find it.

For example, if the target word is "blood" and you say you know "blo" as the first 3 letters, and then you say you know "o" is an extra letter but you don't know where, it could still give you results like "blown" instead of "blood" and "bloom" for example.

But I think this issue is relatively minor in the grand scheme of things for now. So I'll come back to it eventually!