Finished my unit testing!
Like with any good unit testing, it causes you to rethink what your code is doing, so I made a few minor changes to the flow of  how the code runs. Removing some unnecessary redundancies and over-engineered bits that make no sense to me.

Finally the unit tests are done and I'm so happy with it. :)