Over engineering the hell out of my code - big changes
Added a while loop to the main function so you can search again instead of having to restart the whole program after another word guess. Hopefully should help the work flow a bit better.

Also changed the syntax of some of the questions asked in the input statements.

Mainly my biggest change was the over engineering (and maybe messy code additions) of the while loops in the create good and bad lists functions.

Verified that it passed the unittests (had to change my unit test for the new print format. but otherwise it passes.